Minor code cleanup.

+++ Task3/arkpz-pzpi-22-6-kryvko-roman-task3/Atark/src/main/java/ua/nure/kryvko/roman/Atark/greenhouse/GreenhouseService.java
@@ -1,7 +1,10 @@
 package ua.nure.kryvko.roman.Atark.greenhouse;
 
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import ua.nure.kryvko.roman.Atark.user.User;
 import ua.nure.kryvko.roman.Atark.user.UserRepository;
 
@@ -19,6 +22,7 @@ public class GreenhouseService {
         this.userRepository = userRepository;
     }
 
+    @Transactional
     public Greenhouse saveGreenhouse(Greenhouse greenhouse) {
         User owner = userRepository.findById(greenhouse.getUser().getId())
                 .orElseThrow(() -> new IllegalArgumentException("User not found"));
@@ -32,9 +36,11 @@ public class GreenhouseService {
         return greenhouseRepository.findById(id);
     }
 
-    public Greenhouse updateGreenhouse(Greenhouse greenhouse) {
-        if (greenhouse.getId() == null || !greenhouseRepository.existsById(greenhouse.getId())) {
-            throw new IllegalArgumentException("Greenhouse ID not found for update.");
+    @Transactional
+    public Greenhouse updateGreenhouse(Integer id, Greenhouse greenhouse) {
+        greenhouse.setId(id);
+        if (greenhouse.getId() == null || !greenhouseRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Greenhouse ID not found for update.");
         }
 
         User owner = userRepository.findById(greenhouse.getUser().getId())
@@ -44,7 +50,11 @@ public class GreenhouseService {
         return greenhouseRepository.save(greenhouse);
     }
 
+    @Transactional
     public void deleteGreenhouseById(Integer id) {
+        if (!greenhouseRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Greenhouse ID not found for deletion.");
+        }
         greenhouseRepository.deleteById(id);
     }
 }
